test: rework get_mock_paths to be a pytest.fixture
